feat: FeedDetailContent 컴포넌트 구현
