+++ frontend/.storybook/preview.js
@@ -4,6 +4,7 @@ import { QueryClient, QueryClientProvider } from 'react-query';
 
 import GlobalStyle from '../src/Global.styles';
 import ModalProvider from '../src/context/modal/ModalProvider';
+import AsyncBoundary from '../src/components/AsyncBoundary';
 
 export const parameters = {
   actions: { argTypesRegex: '^on[A-Z].*' },
@@ -15,14 +16,23 @@ export const parameters = {
   },
 };
 
-const queryClient = new QueryClient();
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      suspense: true,
+      useErrorBoundary: true,
+    },
+  },
+});
 
 addDecorator((story) => (
   <>
     <QueryClientProvider client={queryClient}>
       <MemoryRouter initialEntries={['/']}>
         <GlobalStyle />
-        <ModalProvider>{story()}</ModalProvider>
+        <AsyncBoundary rejectedFallback={<div>에러났어용 🚨</div>}>
+          <ModalProvider>{story()}</ModalProvider>
+        </AsyncBoundary>
       </MemoryRouter>
     </QueryClientProvider>
   </>
