+++ frontend/src/components/FeedDetailContent/FeedDetailContent.stories.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import FeedDetailContent from './FeedDetailContent';
+
+export default {
+  title: 'components/FeedDetailContent',
+  component: FeedDetailContent,
+  argTypes: {},
+};
+
+const Template = (args: { id: number }) => <FeedDetailContent {...args} />;
+
+export const Default = Template.bind({});
+
+Default.args = {
+  id: 4,
+};
